+++ ChangeLog.md
@@ -6,6 +6,8 @@
 
 - Added support for generating and storing client certificates.
   ([#3](https://github.com/davep/wasat/pull/3))
+- Added support for handling Gemini Protocol input requests (status codes 10
+  and 11) in the CLI. ([#4](https://github.com/davep/wasat/pull/4))
 
 ## v0.0.1